Refactor all methods to their respective files
