+++ src/knarrow/angle_method.pyi
@@ -1,0 +1,6 @@
+from typing import Any
+
+import numpy as np
+import numpy.typing as npt
+
+def angle(x: npt.NDArray[np.float_], y: npt.NDArray[np.float_], kwargs: Any) -> int: ...
