+++ src/knarrow/menger.pyi
@@ -1,0 +1,10 @@
+from typing import Any
+
+import numpy as np
+import numpy.typing as npt
+
+def menger_successive(x: npt.NDArray[np.float_], y: npt.NDArray[np.float_], kwargs: Any) -> int: ...
+def menger_anchored(x: npt.NDArray[np.float_], y: npt.NDArray[np.float_], kwargs: Any) -> int: ...
+def double_triangle_area(vertices: npt.NDArray[np.float_]) -> npt.NDArray[np.float_]: ...
+def get_squared_vector_lengths(vertices: npt.NDArray[np.float_]) -> npt.NDArray[np.float_]: ...
+def get_curvature(vertices: npt.NDArray[np.float_]) -> npt.NDArray[np.float_]: ...
